refactor(devtokens): move token value objects to plugin
